Правки в валидации хэштегов
